Adds modals

+++ src/pages/Dashboard/components/RegistrationCard/registration-card.tsx
@@ -4,27 +4,43 @@ import { ButtonSmall } from '~/components'
 import { DashContext } from '~/context'
 import { handleApi } from '~/utils'
 import * as S from "./styles"
+import { useModal } from '~/hooks'
 
 export const RegistrationCard: FC<RegistrationType> = registration => {
   const { getAll, updateStatus, deleteItem } = handleApi
-  const [, dispatch] = useContext(DashContext)
+  const [{ isFetching }, dispatch] = useContext(DashContext)
+  const handleModal = useModal()
 
-  const handleAction = async (newStatus: RegistrationStatuses) => {
-    dispatch({ type: 'FETCH_CARDS_INIT' })
-    await updateStatus(registration.id, { ...registration, status: newStatus })
-    handleResult()
-  }
+  const handleAction = async (newStatus: RegistrationStatuses | 'DELETE') => {
+    const statusMap: { [K in RegistrationStatuses]: string } = {
+      'REVIEW': 'pronto para revisar',
+      'APROVED': 'aprovado',
+      'REPROVED': 'reprovado'
+    }
 
-  const handleDelete = async () => {
-    dispatch({ type: 'FETCH_CARDS_INIT' })
-    await deleteItem(registration.id)
-    handleResult()
-  }
-  
-  const handleResult = async () => {
-    const { data, error } = await getAll()
-    if (data) dispatch({ type: 'FETCH_CARDS_SUCCESS', payload: data })
-    if (error) dispatch({ type: 'FETCH_CARDS_ERROR', payload: error })
+    const title = newStatus === 'DELETE' ? 'Excluir registro' : 'Alteração de status'
+    const message = newStatus === 'DELETE'
+      ? `Excluindo o registro de ${registration.employeeName}...`
+      : `Alterando o registro de ${registration.employeeName} de ${statusMap[registration.status]} para ${statusMap[newStatus]}...`
+
+    handleModal.open({
+      title,
+      message,
+      onCancel: () => handleModal.close(),
+      onConfirm: async () => {
+        handleModal.disable()
+        dispatch({ type: 'FETCH_CARDS_INIT' })
+        if (newStatus === 'DELETE')
+          await deleteItem(registration.id)
+        else
+          await updateStatus(registration.id, { ...registration, status: newStatus })
+        handleModal.close()
+
+        const { data, error } = await getAll()
+        if (data) dispatch({ type: 'FETCH_CARDS_SUCCESS', payload: data })
+        if (error) dispatch({ type: 'FETCH_CARDS_ERROR', payload: error })
+      }
+    })
   }
 
   return (
@@ -44,16 +60,16 @@ export const RegistrationCard: FC<RegistrationType> = registration => {
       <S.Actions>
         {registration.status === 'REVIEW' && (
           <>
-            <ButtonSmall bgcolor="rgb(255, 145, 154)" onClick={() => handleAction('REPROVED')}>Reprovar</ButtonSmall>
-            <ButtonSmall bgcolor="rgb(155, 229, 155)" onClick={() => handleAction('APROVED')}>Aprovar</ButtonSmall>
+            <ButtonSmall disabled={isFetching} bgcolor="rgb(255, 145, 154)" onClick={() => !isFetching && handleAction('REPROVED')}>Reprovar</ButtonSmall>
+            <ButtonSmall disabled={isFetching} bgcolor="rgb(155, 229, 155)" onClick={() => !isFetching && handleAction('APROVED')}>Aprovar</ButtonSmall>
           </>
         )}
         {registration.status !== 'REVIEW' && (
           <>
-            <ButtonSmall bgcolor="#ff8858" onClick={() => handleAction('REVIEW')}>Revisar novamente</ButtonSmall>
+            <ButtonSmall disabled={isFetching} bgcolor="#ff8858" onClick={() => !isFetching && handleAction('REVIEW')}>Revisar novamente</ButtonSmall>
           </>
         )}
-        <HiOutlineTrash aria-label='delete' onClick={handleDelete} />
+        <HiOutlineTrash aria-label='delete' onClick={() => !isFetching && handleAction('DELETE')} />
       </S.Actions>
     </S.Card>
   )
